fix pipeline circular dep
