+++ python_modules/dagster/dagster_tests/core_tests/system_config_tests/test_system_config.py
@@ -21,6 +21,7 @@ from dagster import (
     types,
 )
 
+
 from dagster.core.system_config.objects import (
     ContextConfig,
     EnvironmentConfig,
@@ -28,7 +29,7 @@ from dagster.core.system_config.objects import (
     SolidConfig,
 )
 
-from dagster.core.system_config.types import (
+from dagster.core.definitions.environment_configs import (
     construct_context_config,
     construct_environment_config,
     construct_solid_dictionary,
@@ -41,6 +42,23 @@ from dagster.core.system_config.types import (
 from dagster.core.types.evaluator import evaluate_config_value
 
 from dagster.core.test_utils import throwing_evaluate_config_value
+
+
+class EnvironmentClassCreationData:
+    def __init__(self, pipeline_name, solids, context_definitions, dependency_structure):
+        self.pipeline_name = pipeline_name
+        self.solids = solids
+        self.context_definitions = context_definitions
+        self.dependency_structure = dependency_structure
+
+
+def create_creation_data(pipeline_def):
+    return EnvironmentClassCreationData(
+        pipeline_def.name,
+        list(pipeline_def._solid_dict.values()),
+        pipeline_def.context_definitions,
+        pipeline_def.dependency_structure,
+    )
 
 
 def test_context_config_any():
@@ -270,7 +288,9 @@ def test_expectations_config():
 def test_solid_dictionary_type():
     pipeline_def = define_test_solids_config_pipeline()
 
-    solid_dict_type = define_solid_dictionary_cls('foobar', pipeline_def).inst()
+    solid_dict_type = define_solid_dictionary_cls(
+        'foobar', create_creation_data(pipeline_def)
+    ).inst()
 
     value = construct_solid_dictionary(
         throwing_evaluate_config_value(
@@ -353,7 +373,9 @@ def test_solid_dictionary_some_no_config():
         ]
     )
 
-    solid_dict_type = define_solid_dictionary_cls('foobar', pipeline_def).inst()
+    solid_dict_type = define_solid_dictionary_cls(
+        'foobar', create_creation_data(pipeline_def)
+    ).inst()
 
     value = construct_solid_dictionary(
         throwing_evaluate_config_value(solid_dict_type, {'int_config_solid': {'config': 1}})
@@ -414,7 +436,7 @@ def test_whole_environment():
 
 def test_solid_config_error():
     solid_dict_type = define_solid_dictionary_cls(
-        'slkdfjkjdsf', define_test_solids_config_pipeline()
+        'slkdfjkjdsf', create_creation_data(define_test_solids_config_pipeline())
     ).inst()
 
     int_solid_config_type = solid_dict_type.fields['int_config_solid'].config_type
